requirements.txt
google-adk[a2a]
google-genai
python-dotenv
mcp
uvicorn
